Change to int64_t & fully-oblivious

+++ include/bucket.cpp
@@ -1,22 +1,22 @@
 #include "bucket.h"
 #include "common.h"
 
-// int finalFlag = 0;
-int nonEnc;
-int oldK;
-int HEAP_NODE_SIZE;
-int MERGE_BATCH_SIZE;
-int FAN_OUT;
-int BUCKET_SIZE;
+// int64_t finalFlag = 0;
+int64_t nonEnc;
+int64_t oldK;
+int64_t HEAP_NODE_SIZE;
+int64_t MERGE_BATCH_SIZE;
+int64_t FAN_OUT;
+int64_t BUCKET_SIZE;
 std::random_device dev3;
 std::mt19937 rng3(dev3());
 // BUCKET_SORT
-void initMerge(int size) {
+void initMerge(int64_t size) {
   HEAP_NODE_SIZE = size;
   MERGE_BATCH_SIZE = size;
 }
 
-bool isTargetIterK(int randomKey, int iter, int k, int num) {
+bool isTargetIterK(int64_t randomKey, int64_t iter, int64_t k, int64_t num) {
   /*
   while (iter) {
     randomKey = randomKey / oldK;
@@ -27,29 +27,29 @@ bool isTargetIterK(int randomKey, int iter, int k, int num) {
   return (randomKey % k) == num;
 }
 
-void mergeSplitHelper(Bucket_x *inputBuffer, int* numRow1, int* numRow2, int* inputId, int* outputId, int iter, int k, int* bucketAddr, int outputStructureId) {
-  // int batchSize = BUCKET_SIZE; // 8192
+void mergeSplitHelper(Bucket_x *inputBuffer, int64_t* numRow1, int64_t* numRow2, int64_t* inputId, int64_t* outputId, int64_t iter, int64_t k, int64_t* bucketAddr, int64_t outputStructureId) {
+  // int64_t batchSize = BUCKET_SIZE; // 8192
   // TODO: FREE these malloc
   // printf("In mergesplitHep!\n");
   Bucket_x **buf = (Bucket_x**)malloc(k * sizeof(Bucket_x*));
-  for (int i = 0; i < k; ++i) {
+  for (int64_t i = 0; i < k; ++i) {
     buf[i] = (Bucket_x*)malloc(BUCKET_SIZE * sizeof(Bucket_x));
   }
-  // int counter0 = 0, counter1 = 0;
-  int randomKey;
-  int *counter = (int*)malloc(k * sizeof(int));
-  memset(counter, 0, k * sizeof(int));
+  // int64_t counter0 = 0, counter1 = 0;
+  int64_t randomKey;
+  int64_t *counter = (int64_t*)malloc(k * sizeof(int64_t));
+  memset(counter, 0, k * sizeof(int64_t));
   // printf("Before assign bins!\n");
-  for (int i = 0; i < k * BUCKET_SIZE; ++i) {
+  for (int64_t i = 0; i < k * BUCKET_SIZE; ++i) {
     if (inputBuffer[i].x != DUMMY) {
       randomKey = inputBuffer[i].key;
-      for (int j = 0; j < k; ++j) {
+      for (int64_t j = 0; j < k; ++j) {
         if (isTargetIterK(randomKey, iter, k, j)) {
           buf[j][counter[j] % BUCKET_SIZE] = inputBuffer[i];
           counter[j]++;
           // std::cout << "couter j: " << counter[j] << std::endl;
           if (counter[j] % BUCKET_SIZE == 0) {
-            opOneLinearScanBlock(bucketAddr[outputId[j]] +  numRow2[outputId[j]], (int*)buf[j], BUCKET_SIZE, outputStructureId, 1, 0);
+            opOneLinearScanBlock(bucketAddr[outputId[j]] +  numRow2[outputId[j]], (int64_t*)buf[j], BUCKET_SIZE, outputStructureId, 1, 0);
             numRow2[outputId[j]] += BUCKET_SIZE;
           }
         }
@@ -57,11 +57,11 @@ void mergeSplitHelper(Bucket_x *inputBuffer, int* numRow1, int* numRow2, int* in
     }
   }
   // printf("Before final assign!\n");
-  for (int j = 0; j < k; ++j) {
+  for (int64_t j = 0; j < k; ++j) {
     if (numRow2[outputId[j]] + (counter[j] % BUCKET_SIZE) > BUCKET_SIZE) {
       printf("overflow error during merge split!\n");
     }
-    opOneLinearScanBlock(bucketAddr[outputId[j]] + numRow2[outputId[j]], (int*)buf[j], counter[j] % BUCKET_SIZE, outputStructureId, 1, BUCKET_SIZE-numRow2[outputId[j]]-(counter[j] % BUCKET_SIZE));
+    opOneLinearScanBlock(bucketAddr[outputId[j]] + numRow2[outputId[j]], (int64_t*)buf[j], counter[j] % BUCKET_SIZE, outputStructureId, 1, BUCKET_SIZE-numRow2[outputId[j]]-(counter[j] % BUCKET_SIZE));
     numRow2[outputId[j]] += counter[j] % BUCKET_SIZE;
     // padWithDummy(outputStructureId, bucketAddr[outputId[j]], numRow2[outputId[j]], BUCKET_SIZE);
     free(buf[j]);
@@ -71,12 +71,12 @@ void mergeSplitHelper(Bucket_x *inputBuffer, int* numRow1, int* numRow2, int* in
   // printf("Finish mergesplitHep!\n");
 }
 
-void mergeSplit(int inputStructureId, int outputStructureId, int *inputId, int *outputId, int k, int* bucketAddr, int* numRow1, int* numRow2, int iter) {
+void mergeSplit(int64_t inputStructureId, int64_t outputStructureId, int64_t *inputId, int64_t *outputId, int64_t k, int64_t* bucketAddr, int64_t* numRow1, int64_t* numRow2, int64_t iter) {
   // step1. Read k buckets together
   // printf("In mergesplit!\n");
   Bucket_x *inputBuffer = (Bucket_x*)malloc(k * sizeof(Bucket_x) * BUCKET_SIZE);
-  for (int i = 0; i < k; ++i) {
-    opOneLinearScanBlock(bucketAddr[inputId[i]], (int*)(&inputBuffer[i * BUCKET_SIZE]), BUCKET_SIZE, inputStructureId, 0, 0);
+  for (int64_t i = 0; i < k; ++i) {
+    opOneLinearScanBlock(bucketAddr[inputId[i]], (int64_t*)(&inputBuffer[i * BUCKET_SIZE]), BUCKET_SIZE, inputStructureId, 0, 0);
   }
   // step2. process k buckets
   mergeSplitHelper(inputBuffer, numRow1, numRow2, inputId, outputId, iter, k, bucketAddr, outputStructureId);
@@ -84,21 +84,21 @@ void mergeSplit(int inputStructureId, int outputStructureId, int *inputId, int *
   // printf("Finish mergesplit!\n");
 }
 
-void kWayMergeSort(int inputStructureId, int outputStructureId, int* numRow1, int* bucketAddr, int bucketNum) {
-  int mergeSortBatchSize = HEAP_NODE_SIZE; // 256
-  int writeBufferSize = MERGE_BATCH_SIZE; // 8192
-  int numWays = bucketNum;
+void kWayMergeSort(int64_t inputStructureId, int64_t outputStructureId, int64_t* numRow1, int64_t* bucketAddr, int64_t bucketNum) {
+  int64_t mergeSortBatchSize = HEAP_NODE_SIZE; // 256
+  int64_t writeBufferSize = MERGE_BATCH_SIZE; // 8192
+  int64_t numWays = bucketNum;
   printf("mergeSortBatchSize %ld, %ld\n", mergeSortBatchSize, writeBufferSize);
   // HeapNode inputHeapNodeArr[numWays];
   HeapNode *inputHeapNodeArr = (HeapNode*)malloc(numWays * sizeof(HeapNode));
-  int totalCounter = 0;
+  int64_t totalCounter = 0;
   
-  int *readBucketAddr = (int*)malloc(sizeof(int) * numWays);
-  memcpy(readBucketAddr, bucketAddr, sizeof(int) * numWays);
-  int writeBucketAddr = 0;
-  int j = 0;
+  int64_t *readBucketAddr = (int64_t*)malloc(sizeof(int64_t) * numWays);
+  memcpy(readBucketAddr, bucketAddr, sizeof(int64_t) * numWays);
+  int64_t writeBucketAddr = 0;
+  int64_t j = 0;
   printf("Initialize\n");
-  for (int i = 0; i < numWays; ++i) {
+  for (int64_t i = 0; i < numWays; ++i) {
     // TODO: 数据0跳过
     if (numRow1[i] == 0) {
       continue;
@@ -107,14 +107,14 @@ void kWayMergeSort(int inputStructureId, int outputStructureId, int* numRow1, in
     node.data = (Bucket_x*)malloc(mergeSortBatchSize * sizeof(Bucket_x));
     node.bucketIdx = i;
     node.elemIdx = 0;
-    opOneLinearScanBlock(readBucketAddr[i], (int*)node.data, std::min(mergeSortBatchSize, numRow1[i]), inputStructureId, 0, 0);
+    opOneLinearScanBlock(readBucketAddr[i], (int64_t*)node.data, std::min(mergeSortBatchSize, numRow1[i]), inputStructureId, 0, 0);
     inputHeapNodeArr[j++] = node;
     readBucketAddr[i] += std::min(mergeSortBatchSize, numRow1[i]);
   }
   
   Heap heap(inputHeapNodeArr, j, mergeSortBatchSize);
   Bucket_x *writeBuffer = (Bucket_x*)malloc(writeBufferSize * sizeof(Bucket_x));
-  int writeBufferCounter = 0;
+  int64_t writeBufferCounter = 0;
   printf("Start merging\n");
   while (1) {
     HeapNode *temp = heap.getRoot();
@@ -125,7 +125,7 @@ void kWayMergeSort(int inputStructureId, int outputStructureId, int* numRow1, in
     
     if (writeBufferCounter == writeBufferSize) {
       nonEnc = 1;
-      opOneLinearScanBlock(writeBucketAddr, (int*)writeBuffer, writeBufferSize, outputStructureId, 1, 0);
+      opOneLinearScanBlock(writeBucketAddr, (int64_t*)writeBuffer, writeBufferSize, outputStructureId, 1, 0);
       nonEnc = 0; 
       writeBucketAddr += writeBufferSize;
       // numRow2[temp->bucketIdx] += writeBufferSize;
@@ -134,7 +134,7 @@ void kWayMergeSort(int inputStructureId, int outputStructureId, int* numRow1, in
     }
     
     if (temp->elemIdx < numRow1[temp->bucketIdx] && (temp->elemIdx % mergeSortBatchSize) == 0) {
-      opOneLinearScanBlock(readBucketAddr[temp->bucketIdx], (int*)(temp->data), std::min(mergeSortBatchSize, numRow1[temp->bucketIdx]-temp->elemIdx), inputStructureId, 0, 0);
+      opOneLinearScanBlock(readBucketAddr[temp->bucketIdx], (int64_t*)(temp->data), std::min(mergeSortBatchSize, numRow1[temp->bucketIdx]-temp->elemIdx), inputStructureId, 0, 0);
       
       readBucketAddr[temp->bucketIdx] += std::min(mergeSortBatchSize, numRow1[temp->bucketIdx]-temp->elemIdx);
       heap.Heapify(0);
@@ -149,7 +149,7 @@ void kWayMergeSort(int inputStructureId, int outputStructureId, int* numRow1, in
     }
   }
   nonEnc = 1;
-  opOneLinearScanBlock(writeBucketAddr, (int*)writeBuffer, writeBufferCounter, outputStructureId, 1, 0);
+  opOneLinearScanBlock(writeBucketAddr, (int64_t*)writeBuffer, writeBufferCounter, outputStructureId, 1, 0);
   nonEnc = 0;
   // numRow2[0] += writeBufferCounter;
   // TODO: ERROR writeBuffer
@@ -158,24 +158,24 @@ void kWayMergeSort(int inputStructureId, int outputStructureId, int* numRow1, in
   free(inputHeapNodeArr);
 }
 
-void bucketSort(int inputStructureId, int size, int dataStart) {
+void bucketSort(int64_t inputStructureId, int64_t size, int64_t dataStart) {
   Bucket_x *arr = (Bucket_x*)malloc(size * sizeof(Bucket_x));
-  opOneLinearScanBlock(dataStart, (int*)arr, size, inputStructureId, 0, 0);
+  opOneLinearScanBlock(dataStart, (int64_t*)arr, size, inputStructureId, 0, 0);
   quickSort(arr, 0, size - 1);
-  opOneLinearScanBlock(dataStart, (int*)arr, size, inputStructureId, 1, 0);
+  opOneLinearScanBlock(dataStart, (int64_t*)arr, size, inputStructureId, 1, 0);
   free(arr);
 }
 
-// int inputTrustMemory[BLOCK_DATA_SIZE];
-int bucketOSort(int structureId, int size) {
+// int64_t inputTrustMemory[BLOCK_DATA_SIZE];
+int64_t bucketOSort(int64_t structureId, int64_t size) {
   double z1 = 6 * (KAPPA + log(2.0*N));
   double z2 = 6 * (KAPPA + log(2.0*N/z1));
   BUCKET_SIZE = BLOCK_DATA_SIZE * ceil(1.0*z2/BLOCK_DATA_SIZE);
   double thresh = 1.0*M/BUCKET_SIZE;
   FAN_OUT = greatestPowerOfTwoLessThan(thresh);
-  int bucketNum = smallestPowerOfKLargerThan(ceil(2.0 * size / BUCKET_SIZE), 2);
-  int k = FAN_OUT;
-  int mem1 = k * BUCKET_SIZE;
+  int64_t bucketNum = smallestPowerOfKLargerThan(ceil(2.0 * size / BUCKET_SIZE), 2);
+  int64_t k = FAN_OUT;
+  int64_t mem1 = k * BUCKET_SIZE;
   HEAP_NODE_SIZE = fmax(floor(1.0 * M / (bucketNum+1)), 1);
   MERGE_BATCH_SIZE = HEAP_NODE_SIZE;
   printf("Mem1's memory: %ld\n", mem1);
@@ -184,37 +184,37 @@ int bucketOSort(int structureId, int size) {
     printf("Memory exceed\n");
   }
   // TODO: Change to ceil, not interger divide
-  int ranBinAssignIters = ceil(1.0*log2(bucketNum)/log2(k));
-  printf("Iteration times: %d\n", (int)ceil(1.0*log2(bucketNum)/log2(k)));
+  int64_t ranBinAssignIters = ceil(1.0*log2(bucketNum)/log2(k));
+  printf("Iteration times: %d\n", (int64_t)ceil(1.0*log2(bucketNum)/log2(k)));
   srand((unsigned)time(NULL));
   // std::cout << "Iters:" << ranBinAssignIters << std::endl;
-  int *bucketAddr = (int*)malloc(bucketNum * sizeof(int));
-  for (int i = 0; i < bucketNum; ++i) {
+  int64_t *bucketAddr = (int64_t*)malloc(bucketNum * sizeof(int64_t));
+  for (int64_t i = 0; i < bucketNum; ++i) {
     bucketAddr[i] = i * BUCKET_SIZE;
   }
-  int *numRow1 = (int*)malloc(bucketNum * sizeof(int));
-  memset(numRow1, 0, bucketNum * sizeof(int));
-  int *numRow2 = (int*)malloc(bucketNum * sizeof(int));
-  memset(numRow2, 0, bucketNum * sizeof(int));
+  int64_t *numRow1 = (int64_t*)malloc(bucketNum * sizeof(int64_t));
+  memset(numRow1, 0, bucketNum * sizeof(int64_t));
+  int64_t *numRow2 = (int64_t*)malloc(bucketNum * sizeof(int64_t));
+  memset(numRow2, 0, bucketNum * sizeof(int64_t));
   
   Bucket_x *trustedMemory = (Bucket_x*)malloc(BUCKET_SIZE * sizeof(Bucket_x));
-  int *inputTrustMemory = (int*)malloc(BUCKET_SIZE * sizeof(int));
-  int total = 0, readStart = 0;
-  int each;
-  int avg = size / bucketNum;
-  int remainder = size % bucketNum;
-  std::uniform_int_distribution<int> dist{0, bucketNum-1};
+  int64_t *inputTrustMemory = (int64_t*)malloc(BUCKET_SIZE * sizeof(int64_t));
+  int64_t total = 0, readStart = 0;
+  int64_t each;
+  int64_t avg = size / bucketNum;
+  int64_t remainder = size % bucketNum;
+  std::uniform_int_distribution<int64_t> dist{0, bucketNum-1};
   aes_init();
   std::cout << "Before initial\n";
-  for (int i = 0; i < bucketNum; ++i) {
+  for (int64_t i = 0; i < bucketNum; ++i) {
     each = avg + ((i < remainder) ? 1 : 0);
     nonEnc = 1;
     // std::cout << "Read initial\n";
     opOneLinearScanBlock(readStart, inputTrustMemory, each, structureId - 1, 0, 0);
     readStart += each;
-    int randomKey;
-    for (int j = 0; j < each; ++j) {
-      // randomKey = (int)oe_rdrand();
+    int64_t randomKey;
+    for (int64_t j = 0; j < each; ++j) {
+      // randomKey = (int64_t)oe_rdrand();
       // randomKey = rand();
       randomKey = dist(rng3);
       trustedMemory[j].x = inputTrustMemory[j];
@@ -222,23 +222,23 @@ int bucketOSort(int structureId, int size) {
     }
     nonEnc = 0;
     // printf("Write initial %d\n", trustedMemory[0].x);
-    opOneLinearScanBlock(bucketAddr[i], (int*)trustedMemory, each, structureId, 1, BUCKET_SIZE-each);
+    opOneLinearScanBlock(bucketAddr[i], (int64_t*)trustedMemory, each, structureId, 1, BUCKET_SIZE-each);
     numRow1[i] += each;
   }
   free(trustedMemory);
   free(inputTrustMemory);
   std::cout << "After initial\n";
   nonEnc = 0;
-  int *inputId = (int*)malloc(k * sizeof(int));
-  int *outputId = (int*)malloc(k *sizeof(int));
-  int outIdx = 0, expo, tempk, jboundary, jjboundary;
-  int k1 = (int)log2(k);
+  int64_t *inputId = (int64_t*)malloc(k * sizeof(int64_t));
+  int64_t *outputId = (int64_t*)malloc(k *sizeof(int64_t));
+  int64_t outIdx = 0, expo, tempk, jboundary, jjboundary;
+  int64_t k1 = (int64_t)log2(k);
   oldK = k1;
-  int tempk_i, tempk_i1;
+  int64_t tempk_i, tempk_i1;
   printf("Before RBA!\n");
   printf("Total bits: %f, each level solved %lu bits\n", log2(bucketNum), k1);
-  for (int i = 0; i < ranBinAssignIters; ++i) {
-    expo = std::min((int)k1, (int)(log2(bucketNum)-i*k1));
+  for (int64_t i = 0; i < ranBinAssignIters; ++i) {
+    expo = std::min((int64_t)k1, (int64_t)(log2(bucketNum)-i*k1));
     tempk = pow(2, expo);
     tempk_i = pow(tempk, i);
     tempk_i1 = tempk_i * tempk;
@@ -249,29 +249,29 @@ int bucketOSort(int structureId, int size) {
     printf("jboundary: %lu, jjboundary: %lu\n", jboundary, jjboundary);
     // finalFlag = (i != (ranBinAssignIters-1)) ? 0 : 1;
     if (i % 2 == 0) {
-      for (int j = 0; j < jboundary; ++j) {
+      for (int64_t j = 0; j < jboundary; ++j) {
         // pass (i-1) * k^i
         //printf("j: %d\n", j);
-        for (int jj = 0; jj < jjboundary; ++jj) {
+        for (int64_t jj = 0; jj < jjboundary; ++jj) {
           //printf("jj: %d\n", jj);
-          for (int m = 0; m < tempk; ++m) {
+          for (int64_t m = 0; m < tempk; ++m) {
             //printf("j, jj, m: %d, %d, %d\n", j, jj, m);
             inputId[m] = j * tempk_i1+ jj + m * jjboundary;
             outputId[m] = (outIdx * tempk + m) % bucketNum;
             // printf("input, output: %d, %d\n", inputId[m], outputId[m]);
           }/*
-          for (int m = 0; m < tempk; ++m) {
+          for (int64_t m = 0; m < tempk; ++m) {
             printf("input%d: %d\n", m, inputId[m]);
           }
-          for (int m = 0; m < tempk; ++m) {
+          for (int64_t m = 0; m < tempk; ++m) {
             printf("output%d: %d\n", m, outputId[m]);
           }*/
           mergeSplit(structureId, structureId + 1, inputId, outputId, tempk, bucketAddr, numRow1, numRow2, i);
           outIdx ++;
         }
       }
-      int count = 0;
-      for (int n = 0; n < bucketNum; ++n) {
+      int64_t count = 0;
+      for (int64_t n = 0; n < bucketNum; ++n) {
         numRow1[n] = 0;
         count += numRow2[n];
       }
@@ -281,28 +281,28 @@ int bucketOSort(int structureId, int size) {
       outIdx = 0;
       //print(arrayAddr, structureId + 1, bucketNum * BUCKET_SIZE);
     } else {
-      for (int j = 0; j < jboundary; ++j) {
+      for (int64_t j = 0; j < jboundary; ++j) {
         //printf("j: %d\n", j);
-        for (int jj = 0; jj < jjboundary; ++jj) {
+        for (int64_t jj = 0; jj < jjboundary; ++jj) {
           //printf("jj: %d\n", jj);
-          for (int m = 0; m < tempk; ++m) {
+          for (int64_t m = 0; m < tempk; ++m) {
             //printf("j, jj, m: %d, %d, %d\n", j, jj, m);
             inputId[m] = j * tempk_i1+ jj + m * jjboundary;
             outputId[m] = (outIdx * tempk + m) % bucketNum;
             //printf("input, output: %d, %d\n", inputId[m], outputId[m]);
           }/*
-          for (int m = 0; m < tempk; ++m) {
+          for (int64_t m = 0; m < tempk; ++m) {
             printf("input%d: %d\n", m, inputId[m]);
           }
-          for (int m = 0; m < tempk; ++m) {
+          for (int64_t m = 0; m < tempk; ++m) {
             printf("output%d: %d\n", m, outputId[m]);
           }*/
           mergeSplit(structureId + 1, structureId, inputId, outputId, tempk, bucketAddr, numRow2, numRow1, i);
           outIdx ++;
         }
       }
-      int count = 0;
-      for (int n = 0; n < bucketNum; ++n) {
+      int64_t count = 0;
+      for (int64_t n = 0; n < bucketNum; ++n) {
         numRow2[n] = 0;
         count += numRow1[n];
       }
@@ -318,16 +318,16 @@ int bucketOSort(int structureId, int size) {
   }
   free(inputId);
   free(outputId);
-  int resultId = 0;
+  int64_t resultId = 0;
   printf("Finish RBA!\n");
   if (ranBinAssignIters % 2 == 0) {
-    for (int i = 0; i < bucketNum; ++i) {
+    for (int64_t i = 0; i < bucketNum; ++i) {
       bucketSort(structureId, numRow1[i], bucketAddr[i]);
     }
     kWayMergeSort(structureId, structureId + 1, numRow1, bucketAddr, bucketNum);
     resultId = structureId + 1;
   } else {
-    for (int i = 0; i < bucketNum; ++i) {
+    for (int64_t i = 0; i < bucketNum; ++i) {
       bucketSort(structureId + 1, numRow2[i], bucketAddr[i]);
     }
     kWayMergeSort(structureId + 1, structureId, numRow2, bucketAddr, bucketNum);
