+++ include/merge.h
@@ -1,6 +1,8 @@
 #ifndef MERGE_H
 #define MERGE_H
 
-int merge_sort(int inStructureId, int outStructureId);
+#include "common.h"
+
+int64_t merge_sort(int64_t inStructureId, int64_t outStructureId);
 
 #endif // !MERGE_H
